+++ modules/apps/portal/portal-upload-impl/src/main/java/com/liferay/portal/upload/internal/ServletFileUpload.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.upload;
+package com.liferay.portal.upload.internal;
 
 import com.liferay.portal.kernel.upload.FileItem;
 import com.liferay.portal.kernel.upload.UploadException;
