--- portal-impl/src/com/liferay/portal/upload/packageinfo
@@ -1,1 +1,0 @@
-version 5.0.0
