+++ portal-kernel/src/com/liferay/portal/kernel/upload/LiferayServletRequest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.upload;
+package com.liferay.portal.kernel.upload;
 
 import java.io.IOException;
